Corrected getDocs()'s fields behavior (partially)
The MongoDB does return an _id field, if you do not explicitly opt out.

The getDocs()'s fields option behavior checked for `-1` to omit fields.
This seems to be wrong, as the MongoDB itself omits fields by `0`.

Also it determined the applied action by the first field only. This ignored all the other fields, which can be mixed.
This implementation still doesn't fully imitate the MongoDB's tough.

Regards!